\docType{data}
\name{AERS}
\alias{AERS}
\title{French university evaluations}
\description{
Created by the French programme law on research of 2006 and up and running since
March 2007, the AERES (Evaluation Agency for Research and Higher Education) is tasked
with evaluating research and higher education institutions, research organisations, research
units, higher education programs and degrees and with approving their staff evaluation procedures. The evaluation is generally done on an ordinal scale, and the results are available on
the agency website 5 . In the present study, the Bachelor's degrees evaluation of March 2011 for
the 23 universities of the French academies of Bordeaux, Toulouse, Lyon, Montpellier, Greno-
ble are analyzed. The universities have been evaluated through four criteria: programme
leadership (pilot training, PT), educational project (EP), schemes for helping students to
succeed (support success, SS), integration of graduates into the job market and continuation
of chosen studies (employability and further studies, EFS).
}

\examples{
  data(AERS)
}

\keyword{datasets}

